requirements.txt
# Core ML and NLP libraries
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# LangChain for RAG pipeline
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Data processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web scraping and data collection
yfinance>=0.2.0
sec-edgar-downloader>=4.0.0
selenium>=4.15.0

# UI and interface
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0

# Optional: For fine-tuning
peft>=0.6.0
accelerate>=0.24.0
datasets>=2.14.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 